code restructure bootstrap from index.js
